Implemented Entities, started drawing UI Elements

+++ Seeds of Apocrypha/src/Game/Windows/Input.h
@@ -42,6 +42,6 @@ namespace Input {
 	bool BtnPressed(int btn);
 	bool BtnReleased(int btn);
 
-	Vector2i MousePos();
+	Vector2f MousePos();
 };
 }
